Implementation
Implementation of the interruption routine for the capture modules 3 and
4 from PIC16F1825. All tested and working, but when only one signal is
treated, the resolution is better. For exemple, when using only capture3
the reading of 1500us signal gives us 1503us. However, when the two
modules are receiving pulses, the reading gives us 1527. This difference
is caused by the operations involving unsigned long long numbers.